Select: Prevent Space key from scrolling the page
Problem: When the menu is focused and the user presses the Space key,
the menu opens and the page scrolls down.

Expected outcome:
- When the menu is focused, the space key should not scroll the page
- When the menu is focused, the space key should open/close the menu
- When the menu is focused and an item is selected, the selected item
is highlighted and scrolls into view

To resolve this, we prevent the page scrolling with e.preventDefault().
It was important to maintain the character navigation feature so we
added a conditional that checks for when the space key is pressed
and the keyBuffer is empty. When this condition is met, the spaces are
not added to the keyBuffer and it will return false. The false value
is needed to handle the space key navigation event that opens/closes
the menu when the space key is pressed.

Bug: T347662
Change-Id: I7de7ad4e7684d965e22ea760bda65f1f4b5b4e36