styles: Increase CSS specificity of subcomponents
- Document a standard where we always add specificity when styling
a subcomponent.
- Increase the specificity of selectors that have subcomponents.
In most cases, we add the subcomponent's root class name like
`cdx-text-input` to the input of the parent component selector.
The purpose of increasing the specificity is to avoid componenent
styles being overridden when new styles load after the component
has loaded.
- Verify that these changes in Codex work in MediaWiki via use of
extension CodexExample and mw.loader.load() in the browser console.

Bug: T351753
Change-Id: Ica5068e3f1cbec4b2be994385526cf71d3b27348