build: Update custom SVGO plugin to conform to public API
In the course of preparing to migrate the codex-icons package over
to ESM, I noticed some type mismatch errors originating from the
custom "removeSvgTag" plugin we include here.

This change updates the body of this plugin to conform to the
SVGO plugin API described here: https://svgo.dev/docs/plugins-api/.

Specifically, we can't just return a modified XML node tree; we need
to return an object with a "root" property that contains either an
enter or exit method which contains instructions for how to modify
a given AST structure.

The change should not have any impact on the final output.

Bug: T352931
Change-Id: I20e7c7091c535afa84045a4c8bb9ac195868d51a