TextArea: Add CSS-only version
TextArea CSS-only version utilizes Less mixins (via
`css-icon.less`) to set the opacity, size, and background
properties of CSS-only start and end icons.
The CSS-only end icon height gets overriden when we call a
mixin to set the icon size so height is set on the CSS-only
end icon so it's consistent with the Vue icon (see Sandbox
and TextArea demo page).
The TextArea demo page documents the CSS-only version
with code examples of TextArea with icons, disabled,
and error state.

Bug: T346815
Change-Id: Ifee0bd2c3688b168e7ba1d0fcda559a6f2c0fe9f