docs, Field: Flatten fieldset examples and cross-link them
4 of the 7 examples in the Field docs were in `<h4>`s under a single
"Fieldsets" `<h3>`. This made them not appear in the side menu. We can't
change the outline setting to show `<h4>`s in the side menu, because the
tooling around component docs injects its own frontmatter that would
override this.

So instead, flatten the structure of the Field docs, changing all these
`<h4>`s into `<h3>`s and adding the word "fieldset" to their title where
appropriate.

Also add some links where the examples refer to other examples, and make
the text in the basic validation example that talks about validating on
blur link to another example that demonstrates how to do this.

Bug: T353846
Change-Id: Id5b33d927c68995aade8249ddfd03ad19d2a801e