build: Remove `<!-- comments -->` from build output
These comments were being included in the build output with things like
createCommentVNode(" eslint-disable vue/no-v-html "). This is wasteful,
but also causes problems: when the root node of a component is a
comment, directives like v-show don't work.

This regressed between v1.0.1 and v1.1.1, but I don't understand why.
Previously, we didn't have to explicitly specify that we wanted to
remove comments, Vite did it for us. I suspect upgrading
@vitejs/plugin-vue may have caused this, but I haven't confirmed that.

Bug: T354394
Change-Id: I1635d13807dc233ff4efc9f42bd723ed517c3a20