build: Update 'svgo' to latest v3.2.0
This update of 'svgo' dependency is a nice new year's gift for us.
It includes three optimizations on converting path commands, which
- improves closing paths and how we determine if to use absolute or
  relative commands.
- round arc or convert to lines based on the geometric sagitta
- convert cubic Bézier curves to quadratic Bézier curves where possible

Those changes lead to SVG file size reduction of up to 7% per icon file.
To give an overview of the savings, only the component demo SVGs are
reduced from
6,847,378 bytes (7.3 MB on disk) to 5,664,494 bytes (6.1 MB on disk),
the SVG icons from
117,576 bytes (1.1 MB on disk) to 115,221 bytes (1.1 MB on disk).
One example is the 'articleNotFound' icon, which is reduced from
842 bytes to 786 bytes (6.6% reduction before gzip).

Bug: T354525
Change-Id: I8b67d559cd264fe319983e78f919c7c627929307