Accordion: Use <details> element for markup
In preparation for the introduction of a CSS-only version of the
Accordion component, this change migrates the existing Vue-based
component to use <details> and <summary> elements instead of the
existing custom markup. This will make it easier for the same
CSS styles to support no-JS usage (the <details> element is necessary
if we want no-JS clients to have the ability to expand and collapse
sections of content).

This change should not expose any visual changes in the current
Accordion component. Assistive technology users may experience
some changes but they should be minor (and will in fact
amount to the default behavior of whatever browser or device they
are using, since we no longer need to define our own custom ARIA
behavior here).

A small amount of new functionality has been added with this change:
users can now set the "open" boolean attribute on Accordion components
to cause them to initially render in the expanded state.

A follow-up patch will demonstrate how a CSS-only Accordion component
could be used.

Bug: T338184
Change-Id: I1d11d40a0c5bb4daa4ed6adb5ff011104f83eca5