Accordion: Enable and demonstrate CSS-only usage
In addition to providing the usual CSS-only usage examples on the
relevant docs page, this commit also makes some changes to how
Accordion.vue handles the <details> indicator icon.

Previously this was handled using a <cdx-icon> component. CSS-only usage
of Accordion at this point would have required the user to bring along
their own CSS-only Icon element and ensure that it is styled correctly.

This seemed less than ideal, especially since the indicator icon is
always the same; it's not like a button icon where the user may or may
not want an icon and they need to control which one they use.

Instead, the Accordion component styles have been re-worked to basically
embed a CSS-only indicator icon in the *Vue* version of the component,
so that it is also available to CSS-only users.

Bug: T338184
Change-Id: Ie1f0041f62c952364cb873b79945582620ad98ba