build: Pin dependencies to exact versions
Adopt strict versioning for dependencies to avoid unexpected changes.
There's in-production positive precedence for this in for example OOUI
for years.
Tested a number of randomly selected scripts throughout the different
packages with no obvious side-effects.

Upgrade postcss to 8.4.33 everywhere, to avoid TypeScript conflicts
between our version of postcss and the (more recent) one that Vite pulls
in.

Bug: T352585
Change-Id: I181b13b2dec8c4379d18750814c28d726aff3cc0