Accordion: Remove click handler, work around test brokenness
Vue Test Utils doesn't handle `<details>` quite correctly. If you click
on the `<summary>` element, or if you programmatically toggle the value
of the `open` attribute, that correctly changes the visibility of the
content and correctly emits a `toggle` event, but the event doesn't have
the `oldState` and `newState` properties. This causes problems for the
Accordion unit tests.

The current code works around this by attaching a `click` handler on the
`<details>` element in the Accordion component, and having the tests
click it. This patch works around it a different way: by toggling the
`open` attribute, and then emitting another `toggle` event to make up
for the first (incorrect) `toggle` event.

Changes:
- Wrap the code to toggle a `<details>` element in a function, and call
  that in each test that toggles the Accordion
- Remove the now-unused click handler (which was only there to make the
  tests work)
- Remove the unused template ref on the `<details>` element

Bug: T338184
Change-Id: I7162418dfbd9d8f5c1a6e492d71e101ca6c6de40