build, tokens: Add a custom "camelCaseNegative" transformer
Introduces a custom "camelCaseNegative" transformer which works
similar to the built-in "name/cti/camel" transformer provided
by Style Dictionary. The only difference is that this transformer
will take any item in a token's path which starts with the "-"
symbol and replace it with the string "NEGATIVE". The
"change-case" library which Style Dictionary relies on for its
own camel transformer is also introduced as a devDependency.

Typically only one name transformer should be used per platform
in Style Dictionary config. However, the new "camelCaseNegative"
has just been added to the existing pipeline after
"name/cti/camel" because it uses a custom matcher (targeting
only tokens that contain one or more path items which start with
the "-" symbol).

This change should produce no actual change in our current token
output from the design-tokens package (the offending token was
already omitted from our built files), but it removes an annoying
error message and would make it safe to publish these tokens in
the future if we chose to do so.

Bug: T355274
Change-Id: I9f81f617a8c90ad91d0d5ef780d26eac4ec6c21c