build: Make diff-css.sh show uncommitted changes
diff-css.sh refuses to run when there are uncommitted changes, and exits
with an error status. Because it runs at the very end of "npm test",
this means that "npm test" will fail if one of the builds results in a
change to a versioned file (most commonly it's because the developer
didn't run svgo after modifying/adding an SVG file).

To make these kinds of CI failures easier to understand and debug, have
diff-css.sh print the diff of the uncommitted changes it encountered,
rather than just printing an easy-to-miss one-line message saying there
are uncommitted changes.

Change-Id: Ib4cbdc3613e03d2fc1e308e9df13204e35c1130b