TextInput, Select: reduce base height
This patch fixes TextInput and Select's base height which is
slightly taller than 32px. Reducing the height to 32px will match
the Figma design spec. Both have a min-height of 32px. We reduced
the line-height of Select and TextInput in order to reduce the
base height of the components.

TextInput and Select's height was reduced from 32.84px to 32px in
Codex Demo. Changes to TextInput affects Combobox, Lookup,
SearchInput, and TypeaheadSearch.

Note: One exception is type="date" inputs are taller than
non-date inputs in Chrome and Safari. Reducing `line-height` from
`x-small` to `xx-small`, reduced its height from 34.84px to 34px.
We'll explore date inputs in a future patch.

These changes do not cause a visual regression for
TypeaheadSearch. The height of the TahS remains the same when the
line-height was reduced.

Bug: T341224
Change-Id: I4b49ea67ef47effca29f96d3f216f71b77c818e1