styles, Field: Fix and simplify spacing of help text
- Apply a consistent margin-top above both the help text and
  the validation message
- Remove the padding-bottom below the control that was
  conditionally applied

The result is that there is a consistent margin between the
control, help text, and validation message, or any combination
of those 3 things.

Also, remove the margin-bottom of the text input on larger
screens in the Field demo with two inputs, since it disrupts
the consistent margins of the Feld component.

Bug: T338802
Change-Id: I0d95617d9d832ed1c1d9f98e0ed32f59d228b0c8