build: Respect CODEX_DOC_ROOT when building the sandbox
The CODEX_DOC_ROOT environment variable is used to override the "base"
setting for Vite in cases where we build the Codex docs site or sandbox
and publish it somewhere other than the domain root (like on
doc.wikimedia.org). The Vite config for the sandbox did respect this env
var, but only in the config that's used for dev mode, not in the config
that's used for build mode.

Also respect CODEX_DOC_ROOT when the sandbox is being built, rather than
hard-coding the base URL to /sandbox/.

Bug: T347107
Change-Id: I0f49292a076e243d2a7768cef0c6eb509afdbed9