docs: Rename to Design System Team (singular)
This commit renames the docs to refer to the Design System Team
instead of the Design Systems Team. This is to reflect the fact
that the system is singular in Wikimedia context.

Also rename to "Design System Team" with an uppercase
"Team" to be consistent with the rest of the documentation.
And fix a few links that have been redirected.

Change-Id: I2989056a9d60ad084258925b366ed4b13c3a344e