build: Automatically add all demos/*.html files to the sandbox build
The sandbox linked to demos/buttons.html, and this worked in dev mode,
but not in build mode because buttons.html wasn't listed as a file to
build. To fix this not just for this demo but for new demos going
forward, automatically include all .html files in the demos/ directory
in the sandbox build.

Also remove some stray CSS in Sandbox.vue that doesn't apply to anything
anymore; the table in question was moved from ButtonDemo.vue to
ButtonGrid.vue (which lives on the separate buttons.html page) a while
ago.

Bug: T357185
Change-Id: I5dedd7b94b512d5173d5fd98de28ff5b8114aec4