docs: Style Accordion content image
In I162f76b5a4435fc8ba30207f6341a06fe728ff45, we removed
base and VitePress styles from the demo page. This included
reset styles for `<img>` tags.

This patch sets the `max-width` to `100%` for `img` elements on the
Accordion demo page in order to prevent the image from
overflowing the container.

Change-Id: I47704b247451c9724ec001b6abc54007b53c15db