docs: Style tables in the demo pane
In I162f76b5a4435fc8ba30207f6341a06fe728ff45, we removed
base and VitePress styles from the demo page. This included
basic styles for tables, which are used in a few demos.

This patch adds enables you to allow some basic table styles
in the demo pane, similar to the option for link styles, by
adding a prop to the Wrapper component that, when true, will
add these styles. This is then applied on the Accordion and
Button demo pages, where tables are currently used.

Change-Id: I4c52a42619d995396609ff2d8284b9735b42f695