Icon, styles: Use mask-image for all CSS icons
background-image will not work when the fill color is a CSS
custom property. Instead, we need to use the mask-image trick
that's currently implemented for icons within buttons, so we
can control the icon's color via a background-color rule.

For browsers that don't support mask-image, background-image
is used instead. A combination of filter (to invert a black
icon to white when needed) and opacity (to make the color
closer to color-base) are used to achieve either a light or
dark icon, depending on which has more contrast. New tokens
were added for filter and opacity so they can be overridden
for night mode.

Bug: T356540
Change-Id: Ibfc9b1be40c7e81ae63598b9b38f92aa5507d685