Lookup: Enable an initial list of suggestions
This changes the Lookup component to store the initial state of
the menuItems prop and reference it when deciding whether to
open or close the menu. If an initial list of menuItems is
provided, the input will open on focus and not close when the
input is cleared. This enables the parent component to show
the initial list when there is no input. This could be used
to show a list of suggested options.

Bug: T350946
Change-Id: I9bef03d10a6233bf597cb2d0c0f28bf1b678a086