docs: Add "no results" message to Lookup demo
This fixes a UI issue that you can reproduce by:
- Typing in the input until there are no results
- Blur, then refocus

The menu will expand with nothing in it. This patch fixes
the immediate issue with this demo by adding a "no results"
message, but we should consider a more robust fix for the
initial suggestions use case in the future.

Bug: T350946
Change-Id: Iacbce4695695af6c80bcf2fed63d9503b0059be7