Select: Use background rules for the CSS-only icon
The CSS icon mixin was recently updated to output `mask` rules
instead of `background` to support the use of color variables
in the form of CSS custom properties.

This works for empty spans or other empty elements, but for the
CSS-only select, we're placing the arrow icon within the
`<select>` element itself. Using `mask-image` affects the entire
element, not just the part with the icon.

Instead, we need to use `background` rules. There are downsides:
- Since this solution is for all browsers, we can't just offer
  the two color options we offer as a fallback for browsers that
  don't support mask-image (approximations of color-base and its
  invert), for two reasons. First, we need to be able to show
  the icon in color-disabled for the select's disabled state.
  Second, in the mask-image fallback, we use opacity to
  approximate the colors, but we can't use that on the entire
  `<select>` element for the same reasons that we can't use
  `mask-image` on it.
- Instead, we need to hardcode the hex values of the two
  colors we need for day mode (color-base and color-disabled).
  This is far from ideal.
- This also means the colors won't automatically change in
  night mode. Those hex values would need to be hardcoded, too.

Bug: T358513
Change-Id: Ida03b57ab8371c7943cfa694176edb25b8e3d129