Field: Remove unused status class from CSS-only Field
The documentation for the CSS-only Field component contained
instructions to add the cdx-field--status-error class to the root
cdx-field element, in addition to adding the validation message div and
adding a status class on the input.

However, this cdx-field--status-error class seems to be completely
unused, and it's not generated by the Vue implementation. So let's
remove it.

Also remove the erroneously placed cdx-field__control--has-help-text
class from the validation message example, this appears to have been
copied from the help text example by mistake.

Change-Id: Ia027c2b3988163f2cb148bacfc9bb0ffce44b05c