tokens: Produce CSS vars file with overrides for dark mode
Update the tokens build script to produce a dark mode overrides file
that overrides CSS variables for color-related tokens with their dark
mode values. This "theme-codex-wikimedia-mode-dark.css" file is meant to be
be used together with the "experimental" Less and CSS files. It only
contains tokens whose value is different in dark mode, so it needs to be
layered on top of the main -experimental.css file.

* Separate out the dark-mode specific build using sd.extend()
  after the first build has been run. Provide an alternative
  build configuration that uses "include" for the default token
  values and "source" for the night mode overrides. This new
  build uses the same "stylesheet" platform as the standard one
  but there is only one file, "theme-codex-wikimedia-mode-dark.css".
* Split the 'base' token type into 'base' (for tokens not overridden by
  the active mode) and 'mode' (for overridden tokens)
* Introduce a new matcher to select only tokens which are defined
  in the dark mode file, based on the new 'mode' type; this is used as a
  filter for the new build.

TODO: Right now this uses a dummy modes/dark.json file for demonstration
purposes. This should be replaced with a real value with real values
once T360341 is done.

Bug: T360343
Change-Id: Iddc50c60de9c227ec3a929f391314509764db8cc