tokens: Use the built-in CSS formatter for the experimental build
Our own custom formatter for the experimental CSS build had so much
removed from it that it was just a thin wrapper around Style
Dictionary's built-in CSS formatter, so use the built-in one instead.

This requires moving the experimental warning to a fileHeader, so do the
same thing for the experimental LESS formatter.

Change-Id: I843a2e02dc75742a9eac6e29a62029f136bf0493