tokens: Add Less mixin version of dark mode build
In addition to `theme-codex-wikimedia-mode-dark.css`, build a .less file
with the same name that contains the same CSS variable overrides, but
wrapped in a Less mixin instead of a `:root` block. This provides more
flexibility for downstream users of Codex to e.g. always apply dark mode
based on a selector by doing something like this:

    .always-dark {
        .cdx-mode-dark();
    }

Bug: T360343
Change-Id: I0454a1d6274561caca1e4290fcb039c8a0a7af2e