tokens: Rename experimental output to theme-codex-wikimedia-experimental
We'd eventually like to rename all instances of theme-wikimedia-ui to
theme-codex-wikimedia, but we can't do that without a breaking change.
For now, only do this for the experimental build, and this will give us
an opening to consider what output files we want to have in the long
term as well: when we remove the -experimental suffix, the files will
still have a different name that doesn't conflict with the old ones.

Change-Id: Ided013ed7255a161abc82753d575b25e3aaedf9d