build: Silence nonsensical CSS syntax errors
During the build process, esbuild outputs ~1200 errors that it claims
are CSS syntax errors ("Expected identifier but found whitespace"), but
aren't actually errors. It appears to complain every time an attribute
selector appears inside a :not() with whitespace around it. Configure
esbuild to silence these errors, to make the build process less noisy.

Change-Id: Id0f4ee712631abfe2064ac2d382a75e4e167c994