tokens: Add dark mode decision tokens
Adds the dark mode decision tokens to dark.json which will
override the color decision tokens in application.json.

This patch adds new tokens and modifies existing option tokens.
- New option token `blue800` added to wikimedia-ui.json because
a darker blue was needed for an interaction state in dark mode.
- New decision tokens were added to application.json and
dark.json to ensure the color is white in both light mode and
dark mode. These new tokens include: `color-inverted-fixed`
and `background-color-base-fixed`.
- Modified the hex code value of option tokens: `red800`,
`yellow800` and `green800.
- Separate patch adjusts the colors in component Vue files and
applies the new tokens with a fixed value between light and dark
mode.
- Ensure the Dialog overlay in light mode is 65% of white, and the overlay
in dark mode should be 65% of black.

Bug: T360341
Change-Id: I9f7eaa87951573382e638ff7fce6aec7f4fef1de