build: Wrap sandbox in a separate `<div dir="...">`
This wrapper div was combined with `<div class="cdx-sandbox">`, but that
broke our ability to set direction-sensitive styles on `.cdx-sandbox`,
since those rules get transformed to `[dir] .cdx-sandbox` or
`[dir="rtl"] .cdx-sandbox`, and those selectors don't match when the
cdx-sandbox div is itself the div that has the dir attribute. This isn't
causing problems now, but could in the future, so let's fix it
preemptively to prevent confusion in the future.

Change-Id: I3569d331c0f6b82f80721e6c0c3f901a2df845f1