Checkbox: Allow label to be visually hidden
This is a rare use case, but came up when building out row
selection in the Table component. This patch adds a prop to
Checkbox, `hideLabel`, which will be passed into the internal
Label component to visually hide the label if necessary.

When the label is visually hidden, the only contents of the
root element are absolutely positioned, giving the root element
no width. Min-sizes have been added to the root element to
prevent this.

Change-Id: I881efbc3c8fdcbcf8d0fad65f73d5cdc7152b774