Table: add sort icon
- Add a new prop `allowSort` to the `TableColumn` type which when true
adds the sort icon to the column's `th` element or header cell. The
sorting logic will be added later in a separate patch.
- Add a Table demo to the sandbox page that displays the sort icon
for a column with text align start(default) and text align end.
- Add styling for the CSS selector `cdx-table__cell--align-end` that
displays the icon before the text.
- Add unit testing for the Table with sort icon displayed.
- Ensure the sort icon is size small (16x16). The icon is positioned
at the start or end of the <th> element depending on whether the
column's text align is start or end. When text align is start then the
icon will be positioned at the end of the <th> and vice versa.

Bug: T362709
Change-Id: Ifb22ce350e8a4ec2e8f89379953c2a20e5015930