Table: add single sort
- Added a new sort prop. When sorting is enabled, it determines the sort
order. Sorting includes ascending, descending, or unsorted. The parent
component binds the sort prop via v-model which is demonstrated in the
Table demo page.
- The sorting icon is conditionally rendered based on the sorting order.
- Added a CSS selector `cdx-table__cell--has-sort` to style the
interactive states of a `thead` element that has a sort enabled. These
styles include hover, active and focus states.
- Added Table demo to display the sorting of data based on the sort
order when the user clicks on the `thead` in a sortable column header.
- Added new types: SortIconMap, Sort, and SortOptions.
- Added aria attributes to improve accessibility.
- Added snapshot and unit tests to test when sorting is enabled. The
unit tests check that an emitted event occurred to update the sort
prop when a user clicks on a sortable column. It also ensures that
sort prop was updated to the correct value.
- Applied the `scope` attribute to `th` elements to define the
relation of header cells and data cells.

Bug: T362710
Co-Authored-by: AnneT <atomasevich@wikimedia.org>
Change-Id: I57e3b246a132e76f24c9e9a0e7821b8ff2deb6d1