build, tokens: Remove all logic for the legacy build
Previously, we switched some token units over from ems to rems,
which allowed us to remove any differences between the "legacy"
and non-legacy builds of Codex design tokens and styles.

This change removes all reference to the legacy build from the
design tokens build scripts. A straight copy is performed via
npm package scripts during the build process to continue to provide
a file named "theme-wikimedia-ui-legacy.less" for backwards-compat
purposes, but this will be removed in the near future.

Bug: T360806
Change-Id: I20b6b06386e92783e4211672d1a42fafbcaf8815
Depends-On: Ic114e284d34cc9ea29701085ed766484cbeea3e8