build, tokens: Fold the "experimental" build into the standard build
This change stops the generation of a separate, experimental build
of Codex which maps certain LESS variables onto CSS custom properties
(which are assumed to have already been defined, but fallback values
are provided just in case).

Going forward, the default build of Codex design tokens and styles will
use this approach. The language and configuration about an extra,
"experimental" build has been removed.

With the removal of the "legacy" and "experimental" builds, we have
greatly simplified the number of different artifacts that users need
to keep track of. Here's what Codex produces after this change:

Design Tokens:

- theme-wikimedia-ui.css (all design tokens as CSS vars; not used in MW)
- theme-wikimedia-ui.scss (all design tokens as SASS vars; not used in MW)
- theme-wikimedia-ui.less (used in MW; described above)
- theme-wikimedia-ui.js (potentially used in MW; all design tokens as JS vars)
- theme-wikimedia-ui.json (all tokens data as json; not used in MW)
- theme-wikimedia-ui-root.css (default and light mode color tokens)
- theme-wikimedia-ui-mode-dark.css (dark mode overrides for color tokens)
- theme-wikimedia-ui-mixin-dark.less (LESS mixin for dark mode overrides)

Codex styles: Only the "rtl" styles are still produced as separate variants.

Bug: T363006
Change-Id: I9cb5e03689b8ca65f4d41ca3ab34999edbb26e35