Table: sortable table headers wrap a button
This patch changes the HTML structure of sortable columns. Now
when sorting is enabled in a column, the `<th>` (non-interactive
element) wraps a `<button>` (interactive element). Using buttons
allow users to use the default key navigations (tab and tab + shift)
to navigate between interactive button elements. Another benefit is
it allows users to key "enter" on the focusable element to trigger
sorting.

In addition:
- Interactive styles (hover, active, and focus) are now applied to
the button nested within the table header.
- Removed native browser/user agent button styles like background
color, border, and outline.
- When the `th` element is sortable, the sort class
`.cdx-table__cell--has-sort` removes the padding from the `th`
element. And padding is added to the button in order for the button
to take up the space (width and height) of its parent container.
- Add Table elements to the tab flow via `tabindex` attribute.
- Updated the Table sort demo on the Sandbox page to have sortable
and non-sortable columns.

Bug: T363613
Change-Id: I29e5a8fb5d103087aa9dac6ccabef86a3ca1779f