Tooltip: Introduce WIP component (directive)
Introduces a basic text-only Tooltip to Codex as a WIP
component. Although we refer to Tooltip as a component in
Codex documentation and design resources, in terms of the
Vue.js implementation what we actually have is a custom
directive. Basic usage looks like:

    <!-- inside some Vue component template -->
    <input v-tooltip="'This input has a tooltip'">

    <!-- with placement specified as an argument -->
    <button v-tooltip:left="'...'">Click me!</button>

See the Vue.js custom directive documentation for more
https://vuejs.org/guide/reusability/custom-directives.html#custom-directives

Directives can be applied to other components as well as
to basic elements in a template. However, there are two
things to keep in mind in this case:

1. A custom directive used on a component will always
  apply to that component's root node (similar to how
  attribute fallthrough behavior works)
2. If a custom directive is used on a component with
  multiple root nodes, it will be ignored and a warning
  will be thrown.

    <!-- This tooltip works as expected, appearing and
         disappearing when its element has focus -->
    <input type="text" v-tooltip="'Hello'">

    <!-- This tooltip will not work as expected, because
         the cdx-tooltip component is wrapped in a div -->
    <cdx-text-input v-model="inputValue" v-tooltip="'Yo'" />

Once we have pinned down the desired Tooltip appearance and
behavior, we will probably want to go back and enable tooltips
in certain other Codex components (TextInput, Fields, Icons...)
by adding the directive to the appropriate inner element and
exposing props to let the user configure the tooltip's behavior.
Exporting a public Tooltip directive would still be useful if
downstream users were building their own custom components and
wanted tooltips there.

FloatingUI is used to handle the Tooltip floating behavior,
placement in relation to reference element, etc.

In addition to the Tooltip directive itself, this patch also
includes some additional changes:

* Adds a new "background-color-inverted" decision token with a
  default value of gray800 and a dark mode value of gray100
* Adds a new "size-tooltip" component token with a value of
  "max-content" (this is needed based on how FloatingUI works)
* Adds a tooltip demo to the sandbox page
* Adds a full-page tooltip demo page (similar to the table demo
  page) to the sandbox site
* Adds a (still barebones) Tooltip docs page
* Adds a new TooltipOptions interface to the types file

Bug: T340456
Change-Id: I2c61ce668629549cd7b16da75d5dfe4a1a949178