[DEPRECATING CHANGE] Checkbox, Radio, ToggleSwitch: Always use Label component in CSS-only version
The CSS-only versions of Checkbox, Radio and ToggleSwitch did not always
use a full CSS-only Label component. In simple cases, they used a plain
`<label>` with a `__label` class. This was inconsistent and caused
styling issues in some cases, where the `display: inline-flex;` rule
applied to these plain labels would cause spaces around links to
disappear.

Update the CSS-only component documentation for these components to
always use a full CSS-only Label. Deprecate the plain label usage and
remove it from the documentation, but continue to support it. The code
supporting it is marked with `// DEPRECATED:` comments, so it can easily
be found and removed in a future 2.0 release.

Bug: T353885
Change-Id: I5754c7a8abad98ff7afa0c1f41ede1e9c4a5e40d