docs: Add spacing between paragraphs and lists in custom blocks
This space was removed in all docs content by adding a negative
margin-top to lists, to reduce the space between lists and
paragraphs, which have a large margin-bottom. However, in custom
blocks, p tags have zero margin, so lists appear too close to
them. This patch adds that space back in for custom blocks.

Change-Id: Ifb424c5dad060d658f4586ea3276080ee22b8170