Table: fix caption when column headers with buttons are sortable
This patch ensures that the caption announces the column headers
with buttons are sortable. To do this we add a <span> element inside
the <caption> element that provides additional information when the
columns have sorting enabled.

We added a new prop called `sortCaption` which is the text
inside the new <span>. Since Codex does not have an i18n system (yet),
you can use this prop to pass in translatable strings. Inside the Table
demo page, we added a warning box to alert users to use this new prop.

Bug: T364636
Change-Id: Ieed6dd05526d187af6f696243750db66ff61899b