Table: Update indeterminate behavior of "select all rows" checkbox
This patch updates the logic for "select all rows" checkbox. When
some checkboxes are selected, the "select all rows" checkbox has
an indeterminate state.

Before, the indeterminate state only occured when all checkboxes
are selected and you individually uncheck a checkbox. Now, an
indeterminate state depends on whether one/some checkboxes are
selected.

Add unit test that check for the indeterminate state. The test
ensure that when some checkboxes are selected, the "select all
rows" checkbox is (1) unchecked, and (2) indeterminate. We check
the indeterminate state via the `:indeterminate` CSS pseudo-class.

Note: Indeterminate is not an HTML/DOM attribute.

Bug: T364927
Change-Id: Ic3df06b226f1d0fb50d177e73f118cb1b55ea0d0