docs: Remove double `dir` attribute on sandbox subpages
In f3eac1dd9f59d3dc94d3504b427a48f35224c62f a direction switcher was
added to the subpages of the sandbox. However, that only works correctly
if every element has exactly one ancestor with a `dir` attribute. The
direction switcher adds such an ancestor, but we were also still setting
`dir="ltr"` on the `<html>` element (which was needed previously,
because things also break if there is no ancestor with a `dir` attribute).
Remove this attribute to make RTL styles work correctly.

Change-Id: I2859a988d02bb36196847547b9468bcd49ee96bb