tokens: Move font-family theme tokens to work around SD bug
In the option tokens (wikimedia-ui.json), font-family.sans is an object
containing several tokens. But in the decision tokens
(application.json), font-family.sans is a single token. When these two
are merged together, it triggers a bug in Style Dictionary where keys
that have both a value and subkeys are not handled correctly, see
https://github.com/amzn/style-dictionary/issues/797 .

To work around this bug, un-nest the font-family.sans object, renaming
the nested tokens from font-family-sans.NNN to font-family.sans-NNN. Do
the same with font-family.serif and font-family.monospace.

Change-Id: I1de831d97a3733fa30b30f69c3b85ccee21a856b