tokens: Don't reformat tokens without references
If a token doesn't contain any references, or only references theme
tokens that we can't replace with a variable reference, then don't
modify it at all. Without this, the effects of transforms would be
undone, which would cause problems if we used CSS vars for tokens that
use the absolute/relative unit transforms.

Change-Id: If5b347f3701d5da21ff66658e53d0986761f3bec