ChipInput: Make chip descriptions optional
This prop is required but was never added to the ChipInput demos
on the docs site, so there are a bunch of warnings in the console.

This patch makes the prop (for both ChipInput and InputChip) optional
with an English default, which is how we're handling other a11y
string props temporarily until T345386 is resolved.

Change-Id: I5f332e992c0e6c5cebb4003a19322d9ceb81fffe