styles, Tooltip: add animation and max-width
This patch adds CSS animation and sets a `max-width`.

* The max-width is set to `@size-1600` (16 rem).

* CSS animation: When you hover over the reference element
with the tooltip, the animation is delayed for 250ms,
progresses in a linear manner, and takes 100ms to complete
the animation. The `animation-fill-mode` CSS property is set
to `both` to extend the animation properties in both forwards
and backward direction. The tooltip animates with opacity and
scale - starting at 0 opacity to 1 opacity and 75% scale to
100% scale. The transform origin is determined by the tooltip's
placement (provided by Floating UI's middleware data) and
appears as though the origin scale is the center of the
triggering element.

* Commented out the `animation-fill-mode` CSS property from
the stylelint config `scale-unlimited/declaration-strict-value`
since there is no Codex design token.

Bug: T367101
Co-Authored-by: Eric Gardner <egardner@wikimedia.orgi>
Change-Id: I3dec973bfb3275c2a864a5d910ab8c93ef331e1b