css-icon: Don't crash when `@param-size-icon` is a non-standard value
If you called `.cdx-mixin-css-icon()` with `@param-size-icon` set to a
non-standard value (anything other than the three size-icon tokens), you
would get an error from the Less compiler. This change fixes this
minimally, by adding a simple fallback that sets the min-size to the
same value as the size in that case.

This doesn't fully support custom icon sizes, because it doesn't
set a meaningful min-size or allow the caller to specify one; but it at
least makes using custom icon sizes possible, even though they work
slightly less well than the standard ones.

Bug: T367098
Change-Id: I9f2d4bf960ab0806741e251c6fccd16d0ffa0051