Button, ToggleButton: improve keypress event handling
Previously, we updated the Button and ToggleButton components
to ensure that the "active" styles got applied for the duration
of keyboard events, for a better UX. Browsers do not implement
:active on keyboard events consistently, so we added some
additional behavior in JS to apply an "active" class on
Space/Enter keydown events, and remove said class on the
corresponding keyup events.

However, one thing we overlooked here was the fact that Space and
Enter keypress events do not have identical behaviors in buttons.
Most browsers implement the following behavior on <button> elements:

* Space keypress: a keydown event is fired repeatedly as long
  as the user holds down the key. Then a single keyup event fires
  on release. Finally, the button's onclick event is triggered
  (button behavior is usually bound to this event). If you bound
  the click handler to a counter, it would increment once.
* Enter keypress: a keydown event immediately followed by an
  onclick event is fired in a repeating pattern for as long as the
  key is held down. Finally, a keyup event is fired on release.
  If you bound the click handler to a counter, it would increment
  continuously for as long as the key was held down.

In a situation where a button launches a Dialog on click (a common
use-case), users would see broken behavior when they used the Enter
key (the dialog opens before the keyup event fires, prevents that
event from happening, and the active styling never gets removed).
This would *not* happen if the Dialog was launched with the Space
key, or if the user clicked the button with the mouse.

This change attempts to pave over the inconsistencies in the
underlying browser behavior by setting up dedicated handlers for
keydown and keyup on Space and Enter keys respectively. Importantly,
the keydown handler uses preventDefault() to stop the automatic
repeated firing of click events; instead the click event is emitted
manually in the keyup handler for each key.

Bug: T314438
Change-Id: I113243ac4048d1072104f678611936394a72db00